refactor: Add SidebarTemplate component for managing sidebar content and behavior
