+++ layouts/sidebar.vue
@@ -1,173 +1,25 @@
 <script setup lang="ts">
-import { signOut } from "firebase/auth";
-
-const classroomStore = useClassroomStore();
-const { classrooms } = storeToRefs(classroomStore);
-
-const op = ref();
-const auth = useFirebaseAuth();
-const router = useRouter();
-const isCollapsed = ref(false);
-
-const toggle = (event: MouseEvent) => {
-    op.value.toggle(event);
-};
-
-const user = useCurrentUser();
-
-const handleProfileClick = () => {
-    router.push("/profile");
-    op.value.hide();
-};
-
-const handleSignOut = async () => {
-    try {
-        if (auth) {
-            await signOut(auth);
-            await useAuth().logout();
-        }
-    } catch (error) {
-        console.error("Error signing out:", error);
-    }
+const sidebarStore = useSidebarStore();
+const { isSidebarOpen, isDrawerOpen } = storeToRefs(sidebarStore) as {
+    isSidebarOpen: Ref<boolean>;
+    isDrawerOpen: Ref<boolean>;
 };
 </script>
 
 <template>
-    <div class="sticky top-[10px] h-[calc(100vh-20px)]">
+    <Drawer v-model:visible="isDrawerOpen" header="ClassCraft">
+        <template #container="{ closeCallback }">
+            <div class="pt-6 px-3 overflow-auto">
+                <SidebarTemplate :closeCallback="closeCallback" />
+            </div>
+        </template>
+    </Drawer>
+    <div class="absolute hidden md:block top-0 h-[100vh] z-[100] md:sticky p-2">
         <div
             class="relative flex flex-col overflow-y-auto justify-between h-full screen pb-0 bg-white border basis-64 rounded-3xl overflow-clip duration-500"
-            :class="isCollapsed ? 'w-20 p-1 py-4' : 'w-72 py-6 px-3'"
+            :class="isSidebarOpen ? 'w-72 py-6 px-3' : 'w-20 p-1 py-4'"
         >
-            <div class="flex flex-col gap-9">
-                <div
-                    class="flex gap-2"
-                    :class="isCollapsed ? 'justify-center' : 'justify-between'"
-                >
-                    <h1
-                        v-if="!isCollapsed"
-                        class="inline-flex items-center text-xl text-primary"
-                    >
-                        <i class="pi pi-sparkles mr-2" />
-                        Class<span class="font-bold">Craft</span>
-                    </h1>
-                    <Button
-                        :icon="
-                            isCollapsed
-                                ? 'pi pi-angle-right'
-                                : 'pi pi-angle-left'
-                        "
-                        severity="secondary"
-                        rounded
-                        aria-label="Toggle sidebar"
-                        @click="isCollapsed = !isCollapsed"
-                    />
-                </div>
-                <div>
-                    <SidebarTab
-                        to="/class"
-                        label="Classroom"
-                        icon="pi pi-book"
-                        :isCollapsed="isCollapsed"
-                    />
-                    <SidebarTab
-                        to="/discussion"
-                        label="Request for class"
-                        icon="pi pi-comments"
-                        :isCollapsed="isCollapsed"
-                    />
-                </div>
-                <div
-                    :class="{ 'flex flex-col items-center gap-1': isCollapsed }"
-                >
-                    <Button v-if="isCollapsed" icon="pi pi-plus" text rounded />
-                    <div v-else class="flex justify-between mb-5">
-                        <p class="text-slate-500">Your classes</p>
-                        <nuxt-link
-                            to="/class/new/edit"
-                            class="text-primary-500 flex items-center gap-1"
-                        >
-                            <i class="pi pi-plus" />
-                            <p>Add</p>
-                        </nuxt-link>
-                    </div>
-                    <SidebarTab
-                        v-for="classroom in classrooms"
-                        :key="classroom?.id"
-                        :to="`/class/${classroom?.id}/edit`"
-                        :classroom="classroom"
-                        :isCollapsed="isCollapsed"
-                    />
-                </div>
-            </div>
-            <div
-                class="w-full flex items-center border-t bg-white sticky bottom-0"
-                :class="
-                    isCollapsed ? 'justify-center py-2' : 'justify-between py-4'
-                "
-            >
-                <div class="flex gap-2 items-center">
-                    <Avatar
-                        :image="user?.photoURL || ''"
-                        shape="circle"
-                        size="large"
-                    />
-                    <div v-if="!isCollapsed" class="leading-4 w-40">
-                        <p class="truncate">
-                            {{ user?.displayName }}
-                        </p>
-                        <p class="text-slate-500 text-sm truncate">
-                            {{ user?.email }}
-                        </p>
-                    </div>
-                </div>
-                <Button
-                    v-if="!isCollapsed"
-                    icon="pi pi-ellipsis-h"
-                    severity="secondary"
-                    rounded
-                    outlined
-                    aria-label="Profile"
-                    @click="toggle"
-                />
-                <Popover ref="op" class="pt-4">
-                    <div
-                        class="flex flex-col justify-center items-center gap-4 w-[12rem]"
-                    >
-                        <Image :alt="`${user?.displayName} profile picture`">
-                            <template #image>
-                                <img
-                                    :src="
-                                        user?.photoURL?.replace(
-                                            's96-c',
-                                            's384-c'
-                                        ) || ''
-                                    "
-                                    :alt="`${user?.displayName} profile picture`"
-                                    class="rounded-full w-24 h-24"
-                                />
-                            </template>
-                        </Image>
-                        <div class="text-center">
-                            <h2 class="text-lg font-bold">
-                                {{ user?.displayName }}
-                            </h2>
-                            <p class="text-slate-500">{{ user?.email }}</p>
-                        </div>
-                        <div class="flex flex-col w-full gap-2">
-                            <Button
-                                label="Profile"
-                                severity="secondary"
-                                @click="handleProfileClick"
-                            />
-                            <Button
-                                @click="handleSignOut"
-                                label="Logout"
-                                severity="danger"
-                            />
-                        </div>
-                    </div>
-                </Popover>
-            </div>
+            <SidebarTemplate />
         </div>
     </div>
 </template>
